change handling of 'type' of data repository
type is set in ctor of KVDataRepository, KVRemoteDataRepository, and
KVDMSDataRepository ("local", "remote", "dms", respectively). this is to
ensure use of correct KVAvailableRunsFile plugin, which are only defined
for these 3 types.
this fixes a bug where a IRODSDataRepository had a KVAvailableRunsFile
base object instead of a DMSAvailableRunsFile, because the type "irods"
was not recognised. now the type of IRODS or SRB data repositories is
that set by their parent class, "dms".