further major improvement/bugfix for KVParticle
copying a particle using a pointer to any of its kinematical frames copies
all of its defined frames (plus parameters and groups) so that the copy
can be used in the same way as the original.

in the copy, the kinematic frame of the passed pointer becomes the default
kinematics of the particle (which may not have been the case for the original),
so that the copy behaves in the same way as the original pointer.