[INDRA]: in order for pedestal corrections to be taken into account...
...each pedestal 'calibration' requires an input signal 'R' or 'L' for the
CsI detector in question. These were no longer created by default for
historic INDRA datasets, but created on demanding during analysis of
data in order to give access to the raw data. Now they need to exist
before any data is read, when calibration parameters are set, so that the
pedestal-corrected 'L_PedCor' and 'R_PedCor' signals can be generated.
The existence of these signals is a pre-requisite for the creation of the
'TotLight' signal the first time that GetDetectorSignal("TotLight") is
called for any INDRA CsI detector.